requirements.txt
appdirs==1.4.4
apturl==0.5.2
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Bottleneck==1.3.5
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==2.1.4
et-xmlfile==1.0.1
fasteners==0.18
fonttools==4.46.0
fs==2.4.16
future==1.0.0
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
intelhex==2.3.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
junit-xml==1.9
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mbed-os-tools==1.8.15
mdurl==0.1.2
mercurial==6.7.2
monotonic==1.6
more-itertools==10.2.0
mpmath==0.0.0
netaddr==0.8.0
netifaces==0.11.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4+dfsg
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
prettytable==2.5.0
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
ssh-import-id==5.11
sympy==1.12
systemd-python==235
tables==3.9.2
texttable==1.6.7
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
usb-creator==0.3.16
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.42.0
xdg==5
xkit==0.0.0
zipp==1.0.0
